refactor: break into smaller components 1
