+++ components/compositions/booking-modal/index.tsx
@@ -8,16 +8,12 @@ import {
   DialogDescription,
   DialogTitle,
 } from '@/components/dialog';
-import CustomDatePicker from './date-picker';
-import TimeSlot from './time-slot';
-import Tabs from '@/components/tabs';
-import SelectedTimeSlotsList from './selected-time-slots-list';
 import BookingForm from './booking-form';
 import { X } from 'lucide-react';
 import PaymentMothod from './payment-method';
 import PaymentInfo from './payment-info';
-import { generateTimeSlotsForDate, notify } from '@/lib/utils';
-import { Tab, tabs } from '@/lib/constants';
+import { notify } from '@/lib/utils';
+import BookingStepOne from './step-one';
 
 interface BookingModalProps {
   isOpen: boolean;
@@ -26,7 +22,6 @@ interface BookingModalProps {
 
 const BookingModal: React.FC<BookingModalProps> = ({ isOpen, setIsOpen }) => {
   const [selectedDate, setSelectedDate] = useState<Date | null>(null);
-  const [selectedTab, setSelectedTab] = useState<Tab>(tabs[0]);
   const [selectedTimeSlots, setSelectedTimeSlots] = useState<
     {
       id: number;
@@ -140,33 +135,6 @@ const BookingModal: React.FC<BookingModalProps> = ({ isOpen, setIsOpen }) => {
     }
   };
 
-  const timeSlots = generateTimeSlotsForDate(selectedDate || new Date());
-
-  const timeSlotsIntoPeriods = timeSlots.reduce((acc, slot) => {
-    if (!acc[slot.period]) {
-      acc[slot.period] = [];
-    }
-    acc[slot.period].push(slot);
-    return acc;
-  }, {} as Record<string, any>);
-
-  const handleDateChange = (date: Date | null) => setSelectedDate(date);
-
-  const onSelectTimeSlot = (slot: any) => {
-    if (selectedTimeSlots.map((t) => t.id).includes(slot.id)) {
-      setSelectedTimeSlots((prev) =>
-        prev.filter((time) => time.id !== slot.id)
-      );
-    } else if (selectedTimeSlots.length < 5) {
-      setSelectedTimeSlots((prev) => [...prev, slot]);
-    } else {
-      notify({
-        message: 'You can only book a maximum of 5 slots at a time',
-        type: 'error',
-      });
-    }
-  };
-
   const totalPrice = selectedTimeSlots.reduce(
     (acc, slot) => acc + (slot.price || 0),
     0
@@ -193,54 +161,12 @@ const BookingModal: React.FC<BookingModalProps> = ({ isOpen, setIsOpen }) => {
           </div>
           <DialogTitle>Book the Pitch</DialogTitle>
           {bookingSteps === 1 ? (
-            <>
-              <DialogDescription>
-                Select a date and time slot to book the pitch
-              </DialogDescription>
-              <DialogBody className='flex flex-col lg:flex-row lg:items-start mt-0 lg:mt-3'>
-                <div className='w-full lg:w-3/5 mt-3 flex justify-start items-center sm:items-start flex-col'>
-                  <CustomDatePicker
-                    selectedDate={selectedDate}
-                    handleDateChange={handleDateChange}
-                  />
-                  {selectedDate && (
-                    <div className='w-full flex flex-col mt-6'>
-                      <Tabs
-                        tabs={tabs}
-                        selectedTab={selectedTab}
-                        setSelectedTab={setSelectedTab}
-                      />
-                      <div className='flex flex-wrap gap-2 my-6 justify-center sm:justify-start items-start'>
-                        {timeSlotsIntoPeriods[selectedTab.id].map(
-                          (time: any) => (
-                            <TimeSlot
-                              key={time.id} // Add a unique key for each time slot
-                              time={time.time}
-                              price={time.price}
-                              isBooked={time.isBooked}
-                              onSelect={() => {
-                                onSelectTimeSlot(time);
-                              }}
-                              selected={selectedTimeSlots
-                                .map((t) => t.id)
-                                .includes(time.id)}
-                            />
-                          )
-                        )}
-                      </div>
-                    </div>
-                  )}
-                </div>
-                <div className='w-full lg:w-2/5 text-center sm:text-left flex justify-center flex-col'>
-                  {selectedDate && (
-                    <SelectedTimeSlotsList
-                      selectedTimeSlots={selectedTimeSlots}
-                      setSelectedTimeSlots={setSelectedTimeSlots}
-                    />
-                  )}
-                </div>
-              </DialogBody>
-            </>
+            <BookingStepOne
+              selectedDate={selectedDate}
+              setSelectedDate={setSelectedDate}
+              selectedTimeSlots={selectedTimeSlots}
+              setSelectedTimeSlots={setSelectedTimeSlots}
+            />
           ) : (
             <>
               <DialogDescription>
@@ -286,7 +212,6 @@ const BookingModal: React.FC<BookingModalProps> = ({ isOpen, setIsOpen }) => {
                 setIsOpen(false);
                 setSelectedDate(null);
                 setSelectedTimeSlots([]);
-                setSelectedTab(tabs[0]);
                 setBookingSteps(1);
               }}
             >
